requirements.txt
aiofiles==24.1.0
aiogram==3.17.0
aiogram.contrib==1.1.3
aiohappyeyeballs==2.4.6
aiohttp==3.11.12
aiosignal==1.3.2
annotated-types==0.7.0
attrs==25.1.0
autopep8==2.3.1
blinker==1.9.0
certifi==2024.8.30
charset-normalizer==3.4.0
click==8.1.7
colorama==0.4.6
Flask==3.1.0
frozenlist==1.5.0
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.4
magic-filter==1.0.12
MarkupSafe==3.0.2
multidict==6.1.0
pillow==11.1.0
propcache==0.2.1
pycodestyle==2.12.1
pydantic==2.10.6
pydantic_core==2.27.2
pygame==2.5.2
python-dotenv==1.0.1
pytube==15.0.0
requests==2.32.3
typing_extensions==4.12.2
urllib3==2.2.3
Werkzeug==3.1.3
yarl==1.18.3
yt-dlp==2025.1.26
